js/src/xpconnect/tests/mochitest/test_xow.html

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
    <title>Tests Cross Origin Wrappers</title>
    <script type="text/javascript" src="/MochiKit/packed.js"></script>
    <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
    <script class="testbody" type="text/javascript;version=1.7">
    var answer = [];
    for (let i in location)
        answer.push(i);
    var expected = [
        "hash",
        "host",
        "hostname",
        "href",
        "pathname",
        "port",
        "protocol",
        "search",
        "reload",
        "replace",
        "assign"
    ];

    is(answer.sort().toString(),
       expected.sort().toString(),
       'enumeration does not work');

    var origProto = window.__proto__;
    try {
        window.__proto__ = window;
        ok(false, 'cyclic proto value allowed');
        window.__proto__ = origProto;
    } catch (e) {
        is(e.toString(), 'Error: cyclic __proto__ value',
           'throw the right exception for a cyclic proto');
        is(window.__proto__, origProto, 'reset __proto__ after a cyclic proto');
    }

    try {
        is('PASS', window.eval("'PASS'"), 'window.eval throws an exception');
    } catch (e) {
        ok(false, 'window.eval does not throw an exception');
    }
    </script>
</body>
</html>
